requirements.txt
asyncpg==0.24.0
colorthief==0.2.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py @ git+https://github.com/iDevision/enhanced-discord.py
Pillow==8.3.2
python-dotenv==0.19.0
requests==2.26.0